LPS-204463 Moving to object-test-util
